requirements.txt
anyio==4.12.0
asgiref==3.11.0
certifi==2025.11.12
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
Django==4.2.27
django-prometheus==2.4.1
django-ratelimit==4.1.0
django-redis==6.0.0
docker==7.1.0
gunicorn==23.0.0
h11==0.16.0
httptools==0.7.1
idna==3.11
msgpack==1.1.2
ngrok==1.7.0
packaging==25.0
prometheus_client==0.23.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
pymemcache==4.0.0
pyp==3.0.9
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
sqlparse==0.5.4
stripe==14.1.0
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
stripe
websockets==15.0.1
y-py==0.6.2
ypy==0.2
django-allauth==65.13.1
